==> 2. date_id.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

# This Python 3 environment comes with many helpful analytics libraries installed
# It is defined by the kaggle/python Docker image: https://github.com/kaggle/docker-python
# For example, here's several helpful packages to load

import numpy as np # linear algebra
import pandas as pd # data processing, CSV file I/O (e.g. pd.read_csv)
import seaborn as sns
import matplotlib.pyplot as plt
from sklearn.preprocessing import LabelEncoder
import datetime
sns.set_theme(style='darkgrid')

# Input data files are available in the read-only "../input/" directory
# For example, running this (by clicking run or pressing Shift+Enter) will list all files under the input directory

import os
for dirname, _, filenames in os.walk('/kaggle/input'):
    for filename in filenames:
        print(os.path.join(dirname, filename))

# You can write up to 20GB to the current directory (/kaggle/working/) that gets preserved as output when you create a version using "Save & Run All"
# You can also write temporary files to /kaggle/temp/, but they won't be saved outside of the current session

import pandas as pd # data processing, CSV file I/O (e.g. pd.read_csv)



**Requerimiento:** analizar la situacion actual de las ventas de la compañía por:

* Producto -> Productos más vendidos y productos menos populares
* Tienda -> Precios por tienda (y añadir cantidades)
* Ciudad -> Mirar solo las de NY, Boston, Philly.

sales = pd.read_csv("/content/drive/MyDrive/MASTER EN DATA SCIENCE/Capstone/data_dsmarket/GrupoRetail_1/Proceso Total/BBDD input/item_sales.csv")
calendar = pd.read_csv("/content/drive/MyDrive/MASTER EN DATA SCIENCE/Capstone/data_dsmarket/GrupoRetail_1/Proceso Total/BBDD input/daily_calendar_with_events.csv")

#Creamos la lista de variables idenfiticadoras y variables de medidas.
lista = sales.columns.values.tolist()
lista_1 = lista[0:7]
del lista[0:7]

#Utilizamos la función MELT
sales = sales.melt(id_vars=lista_1, var_name="date", value_name= "units")

sales.head()

sales.shape

calendar["date"] = pd.to_datetime(calendar["date"], format = "%Y-%m-%d")
calendar["week"] = calendar["date"].dt.isocalendar().week.astype(str)
calendar["year"] = calendar["date"].dt.isocalendar().year.astype(str)
calendar["year_week"] = calendar["year"] + "-" + calendar["week"]

sales.rename(columns={"date": "d"}, inplace = True)

sales.shape

sales_calendar = pd.merge(
    left = sales,
    right = calendar,
    how = 'left',
    on = ["d"])

sales_calendar.head()

lista = sales_calendar.columns.values.tolist()
lista.remove("id")
lista.remove("date")
lista.remove("year_week")
lista

sales_calendar.drop(lista, axis=1, inplace=True)

sales_calendar.head()

date_id = sales_calendar.groupby(["year_week", "id"]).first()

date_id.to_csv("/content/drive/MyDrive/MASTER EN DATA SCIENCE/Capstone/data_dsmarket/GrupoRetail_1/Proceso Total/BBDD output/date_id_2.csv")

date_id
