==> 3. MergeDB.ipynb <==
En este notebook se unifican todos. Salida Final

from google.colab import drive
drive.mount('/content/drive')

import pandas as pd

sales_calendar = pd.read_csv("/content/drive/MyDrive/MASTER EN DATA SCIENCE/Capstone/data_dsmarket/GrupoRetail_1/Proceso Total/BBDD output/sales_calendar_1.csv")

sales_calendar.head()

sales = pd.read_csv("/content/drive/MyDrive/MASTER EN DATA SCIENCE/Capstone/data_dsmarket/GrupoRetail_1/Proceso Total/BBDD input/item_sales.csv")

sales.head()

lista = sales.columns.values.tolist()
lista = lista[7:]
lista
sales.drop(lista, axis=1, inplace=True)

sales.shape

sales_calendar.shape

sales_calendar = pd.merge(
    left = sales_calendar,
    right = sales,
    how = "left",
    on = ["id"]
)

sales_calendar.head()

sales.shape

date_id = pd.read_csv("/content/drive/MyDrive/MASTER EN DATA SCIENCE/Capstone/data_dsmarket/GrupoRetail_1/Proceso Total/BBDD output/date_id_2.csv")

date_id.head()

sales_calendar = pd.merge(
    left = sales_calendar,
    right = date_id,
    how = 'left',
    on = ["id", "year_week"])

sales_calendar.head()

prices = pd.read_csv("/content/drive/MyDrive/MASTER EN DATA SCIENCE/Capstone/data_dsmarket/GrupoRetail_1/Proceso Total/BBDD input/item_prices.csv")

prices["yearweek2"] = prices['yearweek'].astype(str)
prices["year"] = prices['yearweek2'].str[:4]
prices["week"] = prices["yearweek2"].str[4:6]
prices["year_week"] = prices["year"] + "-" + prices["week"]

prices.head()

prices["id"] = prices["item"] + "_" + prices["store_code"]

prices.drop(["yearweek", "yearweek2"], axis=1, inplace=True)

prices.head()

prices.drop(["item", "category", "store_code"], axis=1, inplace = True)

sales_calendar_prices = pd.merge(
    left = sales_calendar,
    right = prices,
    how = "left",
    on = ["id", "year_week" ]
)

sales_calendar_prices.drop(["year_week"], axis=1, inplace = True)

sales_calendar_prices.head()

database = sales_calendar_prices.copy()

database["sell_price"] = database["sell_price"].round(2)

database.to_csv("/content/drive/MyDrive/MASTER EN DATA SCIENCE/Capstone/data_dsmarket/GrupoRetail_1/Proceso Total/BBDD output/database_v5.csv", index=False, header=True, decimal=',', sep=';', float_format='%.3f')

gb = database.groupby(["id", "item", "category", "department", "store", "region", "year"]).sum()

gb.drop("sell_price", axis=1, inplace=True)

gb.reset_index(inplace=True)

gb

gb.to_excel("/content/drive/MyDrive/MASTER EN DATA SCIENCE/Capstone/data_dsmarket/GrupoRetail_1/Proceso Total/BBDD output/db_powerbi_3.xlsx",header=True)

gb

database.to_csv("/content/drive/MyDrive/MASTER EN DATA SCIENCE/Capstone/data_dsmarket/GrupoRetail_1/Proceso Total/BBDD output/database_v5_floats.csv")

database

# sales_calendar_prices.to_csv("database_v2.csv")

# database = pd.read_csv(r"C:\Users\eduar\Downloads\dataset_v4.csv",sep=";")

# database.head()

# database["sell_price"] = database["sell_price"].round(2)

# database.to_csv("database_v5.csv")

# database.to_csv("database_v5.csv", index=False, header=True,
#               decimal=',', sep=';', float_format='%.3f')

# database[database["id"] == "ACCESORIES_1_002_BOS_1"]
