==> 5. Final.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
from sklearn.preprocessing import LabelEncoder
import datetime
sns.set_theme(style='darkgrid')

df = pd.read_csv("/content/drive/MyDrive/MASTER EN DATA SCIENCE/Capstone/data_dsmarket/GrupoRetail_1/Proceso Total/BBDD output/database_v5_floats.csv")

df

df.info()

df.shape

df

df.isnull().sum()

for i in ["date"]:
    df[i]=pd.to_datetime(df[i])

df_sin_duplicados = df.copy()

df_sin_duplicados=df.drop_duplicates().copy(deep=True)

df_sin_duplicados.reset_index(drop=True, inplace=True)

df_sin_duplicados.shape

df_sin_duplicados.isnull().sum()

# df_sin_duplicados['sales'].value_counts()

df_sin_duplicados.info()

df_sin_duplicados.isnull().sum()

pd.to_numeric(df_sin_duplicados['sell_price'], errors='coerce')

# len(df_sin_duplicados[df_sin_duplicados['sales'].isnull()])

df_sinnulos = df_sin_duplicados.copy()

df_sinnulos.dropna(inplace=True)

df_sinnulos.isnull().sum()

df_sinnulos.shape

df_sinnulos.reset_index(drop=True, inplace=True)

df_final = df_sinnulos.copy()

df_final['sales'] = df_final['units'] * df_final['sell_price']

df_final['year'] = df_final['year'].apply(int)

df_final.head()

del(df_final['Unnamed: 0'])

df_final['week'] = df_final['week'].apply(int)

df_final['month'] = pd.DatetimeIndex(df_final['date']).month

df_final['day'] = pd.DatetimeIndex(df_final['date']).day

df_final

df_final.info()

df_final.shape

df_final['item'].unique()

df_final.groupby(['item','store','year'])['sales'].sum()

df_final

weekend = pd.read_csv("/content/drive/MyDrive/MASTER EN DATA SCIENCE/Capstone/data_dsmarket/GrupoRetail_1/Proceso Total/BBDD output/database_weekends_events_4.csv")

del(weekend["Unnamed: 0"])

weekend

df_final

df_final = pd.merge(
    left = df_final,
    right = weekend,
    how = 'left',
    on = ["id", "year", "week"])

df_final

del(df_final["units_y"])

df_final["units"] = df_final["units_x"]
del(df_final["units_x"])

df_final = df_final.replace(np.nan, 0)

df_final.isnull().sum()

df_final

df_final["summer_sales"] = np.where( (df_final.month == 6) |  (df_final.month == 7) |  (df_final.month == 8) , df_final.units, 0)

df_final["christmas_sales"] = np.where( (df_final.month == 12) , df_final.units, 0)

df_final_nan = df_final[ (df_final["units"]) == 0]
print(len(df_final))
print(len(df_final_nan))

min_max_dates = df_final_nan.groupby(["item"]).agg({'date' : [np.min, np.max]}).reset_index()

min_max_dates.columns = range(min_max_dates.shape[1])

lista_1 = ["item", "min", "max"]
lista_dates = min_max_dates.columns.tolist()
dict_dates = dict(zip(lista_dates, lista_1))
min_max_dates.rename( columns =dict_dates, inplace = True)

df_final.info()

precio_inicio = df_final.groupby(["item", "date"]).max("sell_price").reset_index()

precio_inicio["min"] = precio_inicio["date"]

precio_inicio = precio_inicio[["item", "min", "sell_price"]]

precios = pd.merge(
    left = min_max_dates,
    right = precio_inicio,
    how = 'left',
    on = ["item", "min"])

precios.rename(columns = {"sell_price" : "precio_inicio" })

precios['exhibition_time'] = (precios['max'] - precios['min']).dt.days

items_sales = df_final.groupby("item").sum()

items_sales

Discount

from datetime import datetime, timedelta

last_week = df_final.groupby(["date", "item"]).min()
last_week.reset_index(inplace = True)

last_week["date_last_week"] = last_week["date"] - timedelta(days = 7 )

lista = last_week.columns.tolist()
lista
lista.remove("date")
lista.remove("item")
lista.remove("date_last_week")
lista.remove("sell_price")
for i in lista:
    del(last_week[i])

discount = df_final.groupby(["date", "item"]).min()
discount.reset_index (inplace = True)

lista = discount.columns.tolist()
lista
lista.remove("date")
lista.remove("item")
lista.remove("sell_price")
for i in lista:
    del(discount[i])

discount.rename(columns = {"date" : "date_last_week" }, inplace=True)

first_discount = pd.merge(
    left = last_week,
    right = discount,
    how = 'left',
    on = ["date_last_week", "item"])

first_discount.rename(columns = {"sell_price_y" : "precio_last_week" }, inplace=True)

first_discount["cambio"] = first_discount["precio_last_week"]- first_discount["sell_price_x"]

first_discount.groupby("cambio").first()
first_discount.reset_index(inplace = True)

first_discount["discount"] = first_discount["precio_last_week"] / first_discount["sell_price_x"]-1

first_discount["discount"] = np.where( (first_discount.discount < -0.05) , first_discount.discount, 0)

first_discount = first_discount[ (first_discount["discount"]) != 0]

first_discount = first_discount.sort_values(by="date").groupby("item").first()

first_discount.reset_index (inplace = True)
lista = ["date", "sell_price_x", "cambio", "index"]
for i in lista:
    del(first_discount[i])

precios_discounts = pd.merge(
    left = precios,
    right = first_discount,
    how = 'left',
    on = ["item"])

precios_discounts['days_until_first_discount'] = (precios_discounts['date_last_week'] - precios_discounts['min']).dt.days

precios_discounts

precio_first = precios_discounts.groupby("item").first("sell.price")

del(precio_first["exhibition_time"])
del(precio_first["precio_last_week"])
del(precio_first["discount"])
del(precio_first["days_until_first_discount"])
precio_first.rename(columns = {"sell_price" : "precio_inicio" }, inplace=True)

dias_fp = pd.merge(
    left = precio_inicio,
    right = precio_first,
    how = 'left',
    on = ["item"])

dias_fp["discount_on_date"] = dias_fp["sell_price"]- dias_fp["precio_inicio"]

dias_fp["full_price_dummy"] = np.where( (dias_fp.discount_on_date == 0) , 1, 0)

fp_days = dias_fp.groupby("item").sum()
fp_days.reset_index(inplace =True)

fp_days

del(fp_days["sell_price"])
del(fp_days["precio_inicio"])
del(fp_days["discount_on_date"])

precios_discount_fp = pd.merge(
    left = precios_discounts,
    right = fp_days,
    how = 'left',
    on = ["item"])

del(precios_discount_fp["date_last_week"])
del(precios_discount_fp["precio_last_week"])
precios_discount_fp.rename(columns = {"sell_price" : "precio_inicio", "min" : "fecha_inicio_venta", "max": "fecha_fin_venta", "discount" : "first_discount"  }, inplace=True)

precios_discount_fp

df_final["sales_2011"] = np.where( (df_final.year == 2011) , df_final.units, 0)
df_final["sales_2012"] = np.where( (df_final.year == 2012) , df_final.units, 0)
df_final["sales_2013"] = np.where( (df_final.year == 2013) , df_final.units, 0)
df_final["sales_2014"] = np.where( (df_final.year == 2014) , df_final.units, 0)
df_final["sales_2015"] = np.where( (df_final.year == 2015) , df_final.units, 0)
df_final["sales_2016"] = np.where( (df_final.year == 2016) , df_final.units, 0)

df_final.info()

df_final.to_csv("/content/drive/MyDrive/MASTER EN DATA SCIENCE/Capstone/data_dsmarket/GrupoRetail_1/Proceso Total/BBDD output/DF_FINAL.csv")  # este archivo va al sales forecast

items_sales = df_final.groupby(["item","category"]).sum()
items_sales.reset_index(inplace = True)
list = ["year", "week", "month", "day", "weekday_int"]
for i in list:
    del(items_sales[i])

items_sales["mean_pvp"] = items_sales["sales"] / items_sales ["units"]

items_sales["weekend_sale_total"] = items_sales["weekend_sales"] / items_sales ["sales"]

items_sales["event_sales_total"] = items_sales["event_sales"] / items_sales ["sales"]

items_sales["summer_sales_total"] = items_sales["summer_sales"] / items_sales ["sales"]

items_sales["christmas_sales_total"] = items_sales["christmas_sales"] / items_sales ["sales"]

items_sales["2015_vs_2014"] = items_sales["sales_2015"] / items_sales ["sales_2014"]-1

items_sales["2015_vs_2013"] = items_sales["sales_2015"] / items_sales ["sales_2014"]-1

items_sales

df_cluster = pd.merge(
    left = items_sales,
    right = precios_discount_fp,
    how = 'left',
    on = ["item"])

df_cluster["full_price_sales"] = df_cluster["precio_inicio"] * df_cluster["units"]
df_cluster["mean_discount"] = df_cluster["sales"] / df_cluster["full_price_sales"]-1

df_cluster

del(df_cluster["sell_price"])
del(df_cluster["weekend_sales"])
del(df_cluster["event_sales"])
del(df_cluster["summer_sales"])
del(df_cluster["sales_2015"])
del(df_cluster["sales_2011"])
del(df_cluster["sales_2014"])
del(df_cluster["sales_2013"])
del(df_cluster["sales_2012"])
del(df_cluster["sales_2016"])
del(df_cluster["fecha_inicio_venta"])
del(df_cluster["fecha_fin_venta"])
del(df_cluster["christmas_sales"])
del(df_cluster["event_dummy"])

df_cluster

df_cluster.to_excel("/content/drive/MyDrive/MASTER EN DATA SCIENCE/Capstone/data_dsmarket/GrupoRetail_1/Proceso Total/BBDD output/bd_clustering.xlsx")

# EL FICHERO bd_clustering SE MODIFICARÁ AÑADIENDO MÁS DATOS PASANDO A SER EL FICHERO bbdd_clustering_v1 QUE SE ENCUENTRA EN EL DIRECTORIO OUTPUT
